requirements.txt
Appium-Python-Client==0.24
beautifulsoup4==4.5.3
bs4==0.0.1
editdistance==0.3.1
lxml==3.7.2
numpy==1.13.1
selenium==3.4.3
SQLAlchemy==1.1.13
terminaltables==3.1.0
